New "Edit item" page

+++ project-frontend/src/components/Admin/AdminItems.tsx
@@ -4,14 +4,9 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import { Category, Item } from '../../types/types';
 import { RootState } from '../../reducers/rootReducer';
-import { UseField } from '../../hooks/useField';
-import { UseTextArea } from '../../hooks/useTextArea';
 
 import { handleError } from '../../util/handleError';
 import itemService from '../../services/itemService';
-import item_categoryService from '../../services/item_categoryService';
-import useField from '../../hooks/useField';
-import useTextArea from '../../hooks/useTextArea';
 import { isNumber } from '../../types/type_functions';
 
 import { setNotification } from '../../reducers/miscReducer';
@@ -19,13 +14,6 @@ import { setNotification } from '../../reducers/miscReducer';
 import AdminItem from './AdminItem';
 import AddItemForm from '../Admin/AddItemForm';
 import { Link } from '../CustomLink';
-
-export interface ItemInputs {
-    name: UseField;
-    description: UseTextArea;
-    price: UseField;
-    instock: UseField;
-}
 
 const AdminItems = () => {
     const dispatch = useDispatch();
@@ -41,16 +29,6 @@ const AdminItems = () => {
 
     const [category, setCategory] = useState<Category | undefined>(getCategoryFromParams());
     const [items, setItems] = useState<Item[]>([]);
-    const [categoriesChanged, setCategoriesChanged] = useState<boolean>(false);
-    const [editedItem, setEditedItem] = useState<Item | null>(null);
-    const [selectedCategories, setSelectedCategories] = useState<number[]>([]);
-
-    const inputs: ItemInputs = {
-        name: useField('text'),
-        description: useTextArea(),
-        price: useField('decimal'),
-        instock: useField('integer'),
-    };
 
     useEffect(() => {
         const id = Number(searchParams.get('category'));
@@ -74,6 +52,7 @@ const AdminItems = () => {
 
     useEffect(() => {
         refreshItems();
+        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [category]);
 
     const deleteItem = async (item: Item) => {
@@ -86,78 +65,6 @@ const AdminItems = () => {
             dispatch(setNotification({ tone: res.success ? 'Positive' : 'Negative', message: res.message }));
 
             refreshItems();
-        }
-    };
-
-    const editItem = (item: Item) => {
-        setEditedItem(item);
-
-        inputs.name.setNewValue(item.name);
-        inputs.description.setNewValue(item.description);
-        inputs.price.setNewValue(item.price.toString());
-        inputs.instock.setNewValue(item.instock.toString());
-    };
-
-    const editItemCancel = () => {
-        setEditedItem(null);
-    };
-
-    const editItemSubmit = async (event: React.FormEvent) => {
-        event.preventDefault();
-        if (
-            editedItem &&
-            (editedItem.name !== inputs.name.value ||
-                editedItem.description !== inputs.description.value ||
-                editedItem.price.toString() !== inputs.price.value.toString() ||
-                editedItem.instock.toString() !== inputs.instock.value.toString() ||
-                categoriesChanged)
-        ) {
-            if (usersState.loggedUser && usersState.loggedUser.token) {
-                const token = usersState.loggedUser.token;
-
-                const updatedItem = {
-                    ...editedItem,
-                    name: inputs.name.value.toString(),
-                    description: inputs.description.value.toString(),
-                    price: Number(inputs.price.value),
-                    instock: Number(inputs.instock.value),
-                };
-
-                // Add connections between the edited Item and the selected Categories that are not yet connected to the Item:
-                selectedCategories.forEach(async (selected) => {
-                    const category = categoryState.find((c) => {
-                        return c.id === selected;
-                    });
-                    if (category && editedItem.categories.includes(category) === false) {
-                        const res = await item_categoryService.addConnection(editedItem, category, token);
-                        if (!res.success) {
-                            handleError(new Error(res.message));
-                        }
-                    }
-                });
-
-                // Remove connections between the edited Item and Categories that are currently connected to the Item:
-                const toRemove = editedItem.categories.filter((c) => {
-                    selectedCategories.includes(c.id) === false;
-                });
-                toRemove.forEach(async (c) => {
-                    const res = await item_categoryService.deleteConnection(editedItem.id, c.id, token);
-                    if (!res.success) {
-                        handleError(new Error(res.message));
-                    }
-                });
-
-                // Update the other info (name, description, etc):
-                const res = await itemService.update(updatedItem, usersState.loggedUser.token, dispatch);
-
-                dispatch(setNotification({ tone: res.success ? 'Positive' : 'Negative', message: res.message }));
-
-                setEditedItem(null);
-
-                refreshItems();
-            } else {
-                handleError(new Error('Missing token'));
-            }
         }
     };
 
@@ -179,40 +86,29 @@ const AdminItems = () => {
             </table>
             <h4>{category ? category.name : 'Uncategorized'}</h4>
             <p>{category ? category.description : 'Items that do not currently belong to any category'}</p>
-            <form onSubmit={editItemSubmit} className='adminFormItemEdit'>
-                <table align='center' width='100%' className='sizeSmallish paddingTopBottomOnly dotted adminItems'>
-                    <tbody>
-                        <tr className='bold'>
-                            <td width='12%' style={{ maxWidth: '25%' }}>
-                                Product
-                            </td>
-                            <td>Description</td>
-                            <td>Price</td>
-                            <td width='1px' className='noWrap'>
-                                In stock
-                            </td>
-                            <td width='1px'>ID</td>
-                            <td width='1px'>Categ.</td>
-                            <td width='1px'>Images</td>
-                            <td width='1px' style={{ paddingRight: 0 }}></td>
-                            <td width='1px'></td>
-                        </tr>
-                        {items.map((item) => (
-                            <AdminItem
-                                key={item.id}
-                                item={item}
-                                isEdited={editedItem === item}
-                                inputs={inputs}
-                                deleteItem={deleteItem}
-                                editItem={editItem}
-                                editItemCancel={editItemCancel}
-                                setSelectedCategories={setSelectedCategories}
-                                setCategoriesChanged={setCategoriesChanged}
-                            />
-                        ))}
-                    </tbody>
-                </table>
-            </form>
+
+            <table align='center' width='100%' className='sizeSmallish paddingTopBottomOnly dotted adminItems'>
+                <tbody>
+                    <tr className='bold'>
+                        <td width='12%' style={{ maxWidth: '25%' }}>
+                            Product
+                        </td>
+                        <td>Description</td>
+                        <td>Price</td>
+                        <td width='1px' className='noWrap'>
+                            In stock
+                        </td>
+                        <td width='1px'>ID</td>
+                        <td width='1px'>Categ.</td>
+                        <td width='1px'>Images</td>
+                        <td width='1px' style={{ paddingRight: 0 }}></td>
+                        <td width='1px'></td>
+                    </tr>
+                    {items.map((item) => (
+                        <AdminItem key={item.id} item={item} deleteItem={deleteItem} />
+                    ))}
+                </tbody>
+            </table>
             <br />
             {usersState.loggedUser?.admin ? <AddItemForm user={usersState.loggedUser} category={category} items={items} setItems={setItems} /> : <></>}
         </div>
